refactor to reduce code redundancy between FakeInsertValuesExecuter and INsertValuesExecuter
