+++ ksql-functional-tests/src/test/java/io/confluent/ksql/test/tools/FakeInsertValuesExecutorTest.java
@@ -1,11 +1,17 @@
 package io.confluent.ksql.test.tools;
 
-import com.google.common.collect.ImmutableList;
 import io.confluent.ksql.engine.KsqlEngine;
 import io.confluent.ksql.function.TestFunctionRegistry;
 import io.confluent.ksql.json.JsonMapper;
-import io.confluent.ksql.parser.tree.*;
+import io.confluent.ksql.parser.KsqlParser;
+import io.confluent.ksql.parser.tree.DoubleLiteral;
+import io.confluent.ksql.parser.tree.Expression;
+import io.confluent.ksql.parser.tree.InsertValues;
+import io.confluent.ksql.parser.tree.IntegerLiteral;
+import io.confluent.ksql.parser.tree.QualifiedName;
+import io.confluent.ksql.parser.tree.StringLiteral;
 import io.confluent.ksql.services.ServiceContext;
+import io.confluent.ksql.statement.ConfiguredStatement;
 import io.confluent.ksql.test.model.QttTestFile;
 import io.confluent.ksql.test.model.TestCaseNode;
 import io.confluent.ksql.test.serde.string.StringSerdeSupplier;
@@ -21,8 +27,10 @@ import org.mockito.junit.MockitoJUnitRunner;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.List;
 import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -72,16 +80,21 @@ public class FakeInsertValuesExecutorTest {
   }
 
   @Test
-  public void shouldExecuteDelimited() {
+  public void shouldExecute() {
     // Given:
+    List<String> columns = new ArrayList<>();
+    columns.add("ID");
+    columns.add("NAME");
+    columns.add("VALUE");
     List<Expression> values = new ArrayList<>();
     values.add(new IntegerLiteral(1034));
     values.add(new StringLiteral("k"));
     values.add(new DoubleLiteral("45.3"));
-    InsertValues insertValues = new InsertValues(QualifiedName.of("TEST"), ImmutableList.of(), values);
-
+    InsertValues insertValues = new InsertValues(QualifiedName.of("TEST"), columns, values);
+    Map<String, Object> overrides = new HashMap<>();
     // When:
-    FakeInsertValuesExecutor.execute(ksqlEngine, ksqlConfig, fakeKafkaService, insertValues);
+    new FakeInsertValuesExecutor(fakeKafkaService)
+            .run(ConfiguredStatement.of(KsqlParser.PreparedStatement.of(insertValues.toString(), insertValues), overrides, ksqlConfig), ksqlEngine, serviceContext);
 
     // Then:
     assertThat(fakeKafkaService.readRecords("test_topic").get(0).getTestRecord().value, equalTo("1034,k,45.3"));
