+++ ksql-functional-tests/src/test/resources/test-runner/test5/output.json
@@ -1,7 +1,7 @@
 {
   "outputs": [
-    {"topic": "S1", "key": 101, "value": "abc,13.54"},
-    {"topic": "S1", "key": 123, "value": "bar,"},
-    {"topic": "S1", "key": 9001, "value": ","}
+    {"topic": "S1", "key": "abc", "value": "abc,13.54"},
+    {"topic": "S1", "key": "bar", "value": "bar,"},
+    {"topic": "S1", "key": "nar", "value": "nar,"}
   ]
 }
