+++ ksql-functional-tests/src/test/resources/test-runner/test5/statements.sql
@@ -1,6 +1,6 @@
-CREATE STREAM TEST (ID int, NAME varchar, VALUE double) WITH (kafka_topic='test_topic', value_format='DELIMITED', key='ID');
-INSERT INTO TEST VALUES (101, 'abc', 13.54);
-INSERT INTO TEST VALUES (30, 'foo', 4.5);
+CREATE STREAM TEST (ID int, NAME varchar, VALUE double) WITH (kafka_topic='test_topic', value_format='DELIMITED', key='NAME');
+INSERT INTO TEST VALUES ('abc', 101, 'abc', 13.54);
+INSERT INTO TEST VALUES ('foo', 30, 'foo', 4.5);
 INSERT INTO TEST (ID, NAME) VALUES (123, 'bar');
-INSERT INTO TEST (ID) VALUES (9001);
+INSERT INTO TEST (ROWKEY, ID) VALUES ('nar', 9001);
 CREATE STREAM S1 as SELECT name, value FROM test where id > 100;
