+++ ksql-functional-tests/src/test/resources/test-runner/test6/output.json
@@ -1,7 +1,7 @@
 {
   "outputs": [
-    {"topic": "S1", "key": 101, "value": {"NAME": "abc", "VALUE": 13.54}},
-    {"topic": "S1", "key": 123, "value": {"NAME": "bar", "VALUE": null}},
-    {"topic": "S1", "key": 43245, "value": {"NAME": "far", "VALUE": 43.0}}
+    {"topic": "S1", "key": "abc", "value": {"NAME": "abc", "VALUE": 13.54}},
+    {"topic": "S1", "key": "bar", "value": {"NAME": "bar", "VALUE": null}},
+    {"topic": "S1", "key": "far", "value": {"NAME": "far", "VALUE": 43.0}}
   ]
 }
