+++ logs/ksql-server-gc.log.0.current
@@ -1,0 +1,36 @@
+Java HotSpot(TM) 64-Bit Server VM (25.211-b12) for bsd-amd64 JRE (1.8.0_211-b12), built on Apr  1 2019 20:53:18 by "java_re" with gcc 4.2.1 (Based on Apple Inc. build 5658) (LLVM build 2336.11.00)
+Memory: 4k page, physical 16777216k(417364k free)
+
+/proc/meminfo:
+
+CommandLine flags: -XX:+CMSClassUnloadingEnabled -XX:+CMSScavengeBeforeRemark -XX:+ExplicitGCInvokesConcurrent -XX:GCLogFileSize=104857600 -XX:InitialHeapSize=268435456 -XX:+ManagementServer -XX:MaxHeapSize=3221225472 -XX:MaxTenuringThreshold=6 -XX:NewRatio=1 -XX:NumberOfGCLogFiles=10 -XX:OldPLABSize=16 -XX:+PrintGC -XX:+PrintGCDateStamps -XX:+PrintGCDetails -XX:+PrintGCTimeStamps -XX:+UseCompressedClassPointers -XX:+UseCompressedOops -XX:+UseConcMarkSweepGC -XX:+UseGCLogFileRotation -XX:+UseParNewGC 
+2019-06-14T15:44:34.067+0800: 0.962: [GC (Allocation Failure) 2019-06-14T15:44:34.067+0800: 0.962: [ParNew: 104960K->12031K(118016K), 0.0108261 secs] 104960K->12031K(249088K), 0.0109373 secs] [Times: user=0.02 sys=0.00, real=0.01 secs] 
+2019-06-14T15:44:34.941+0800: 1.836: [GC (Allocation Failure) 2019-06-14T15:44:34.941+0800: 1.836: [ParNew: 116991K->9078K(118016K), 0.0269747 secs] 116991K->13807K(249088K), 0.0270647 secs] [Times: user=0.07 sys=0.01, real=0.02 secs] 
+2019-06-14T15:44:34.968+0800: 1.863: [GC (CMS Initial Mark) [1 CMS-initial-mark: 4729K(131072K)] 13807K(249088K), 0.0010901 secs] [Times: user=0.00 sys=0.00, real=0.00 secs] 
+2019-06-14T15:44:34.969+0800: 1.865: [CMS-concurrent-mark-start]
+2019-06-14T15:44:34.997+0800: 1.893: [CMS-concurrent-mark: 0.028/0.028 secs] [Times: user=0.08 sys=0.01, real=0.03 secs] 
+2019-06-14T15:44:34.997+0800: 1.893: [CMS-concurrent-preclean-start]
+2019-06-14T15:44:34.998+0800: 1.894: [CMS-concurrent-preclean: 0.001/0.001 secs] [Times: user=0.01 sys=0.00, real=0.00 secs] 
+2019-06-14T15:44:34.999+0800: 1.894: [GC (CMS Final Remark) [YG occupancy: 10879 K (118016 K)]2019-06-14T15:44:34.999+0800: 1.894: [GC (CMS Final Remark) 2019-06-14T15:44:34.999+0800: 1.894: [ParNew: 10879K->2052K(118016K), 0.0046492 secs] 15608K->9030K(249088K), 0.0047124 secs] [Times: user=0.01 sys=0.00, real=0.01 secs] 
+2019-06-14T15:44:35.003+0800: 1.899: [Rescan (parallel) , 0.0015884 secs]2019-06-14T15:44:35.005+0800: 1.901: [weak refs processing, 0.0001062 secs]2019-06-14T15:44:35.005+0800: 1.901: [class unloading, 0.0039685 secs]2019-06-14T15:44:35.009+0800: 1.905: [scrub symbol table, 0.0021099 secs]2019-06-14T15:44:35.011+0800: 1.907: [scrub string table, 0.0002412 secs][1 CMS-remark: 6978K(131072K)] 9030K(249088K), 0.0135005 secs] [Times: user=0.04 sys=0.00, real=0.02 secs] 
+2019-06-14T15:44:35.012+0800: 1.908: [CMS-concurrent-sweep-start]
+2019-06-14T15:44:35.015+0800: 1.911: [CMS-concurrent-sweep: 0.003/0.003 secs] [Times: user=0.01 sys=0.00, real=0.00 secs] 
+2019-06-14T15:44:35.016+0800: 1.911: [CMS-concurrent-reset-start]
+2019-06-14T15:44:35.018+0800: 1.914: [CMS-concurrent-reset: 0.002/0.002 secs] [Times: user=0.01 sys=0.00, real=0.00 secs] 
+2019-06-14T15:44:35.821+0800: 2.716: [GC (Allocation Failure) 2019-06-14T15:44:35.821+0800: 2.716: [ParNew: 107012K->9106K(118016K), 0.0058707 secs] 113985K->16080K(249088K), 0.0059457 secs] [Times: user=0.02 sys=0.00, real=0.00 secs] 
+2019-06-14T15:44:36.712+0800: 3.608: [GC (Allocation Failure) 2019-06-14T15:44:36.712+0800: 3.608: [ParNew: 114066K->8193K(118016K), 0.0110577 secs] 121040K->19219K(249088K), 0.0111547 secs] [Times: user=0.04 sys=0.00, real=0.01 secs] 
+2019-06-14T15:44:37.809+0800: 4.704: [GC (Allocation Failure) 2019-06-14T15:44:37.809+0800: 4.704: [ParNew: 113153K->6602K(118016K), 0.0111774 secs] 124179K->20092K(249088K), 0.0112722 secs] [Times: user=0.02 sys=0.00, real=0.02 secs] 
+2019-06-14T15:44:37.820+0800: 4.716: [GC (CMS Initial Mark) [1 CMS-initial-mark: 13490K(131072K)] 20097K(249088K), 0.0018105 secs] [Times: user=0.01 sys=0.00, real=0.00 secs] 
+2019-06-14T15:44:37.822+0800: 4.718: [CMS-concurrent-mark-start]
+2019-06-14T15:44:37.879+0800: 4.775: [CMS-concurrent-mark: 0.057/0.057 secs] [Times: user=0.06 sys=0.00, real=0.05 secs] 
+2019-06-14T15:44:37.879+0800: 4.775: [CMS-concurrent-preclean-start]
+2019-06-14T15:44:37.881+0800: 4.777: [CMS-concurrent-preclean: 0.002/0.002 secs] [Times: user=0.00 sys=0.00, real=0.01 secs] 
+2019-06-14T15:44:37.899+0800: 4.794: [GC (CMS Final Remark) [YG occupancy: 8581 K (118016 K)]2019-06-14T15:44:37.906+0800: 4.802: [GC (CMS Final Remark) 2019-06-14T15:44:37.906+0800: 4.802: [ParNew: 8581K->4206K(118016K), 0.0029421 secs] 22072K->17696K(249088K), 0.0030237 secs] [Times: user=0.01 sys=0.00, real=0.00 secs] 
+2019-06-14T15:44:37.909+0800: 4.805: [Rescan (parallel) , 0.0125475 secs]2019-06-14T15:44:37.922+0800: 4.817: [weak refs processing, 0.0008129 secs]2019-06-14T15:44:37.923+0800: 4.818: [class unloading, 0.0072412 secs]2019-06-14T15:44:37.930+0800: 4.825: [scrub symbol table, 0.0080198 secs]2019-06-14T15:44:37.938+0800: 4.833: [scrub string table, 0.0005863 secs][1 CMS-remark: 13490K(131072K)] 17696K(249088K), 0.0402726 secs] [Times: user=0.04 sys=0.00, real=0.04 secs] 
+2019-06-14T15:44:37.939+0800: 4.835: [CMS-concurrent-sweep-start]
+2019-06-14T15:44:37.946+0800: 4.841: [CMS-concurrent-sweep: 0.006/0.006 secs] [Times: user=0.02 sys=0.00, real=0.01 secs] 
+2019-06-14T15:44:37.946+0800: 4.841: [CMS-concurrent-reset-start]
+2019-06-14T15:44:37.947+0800: 4.842: [CMS-concurrent-reset: 0.001/0.001 secs] [Times: user=0.00 sys=0.00, real=0.00 secs] 
+2019-06-14T15:44:55.149+0800: 22.044: [GC (Allocation Failure) 2019-06-14T15:44:55.149+0800: 22.044: [ParNew: 109166K->13055K(118016K), 0.0109916 secs] 122327K->27880K(249088K), 0.0112241 secs] [Times: user=0.03 sys=0.00, real=0.02 secs] 
+2019-06-14T16:07:54.609+0800: 1401.492: [GC (Allocation Failure) 2019-06-14T16:07:54.612+0800: 1401.495: [ParNew: 118015K->6948K(118016K), 0.0185176 secs] 132840K->28797K(249088K), 0.0211650 secs] [Times: user=0.04 sys=0.02, real=0.03 secs] 
+2019-06-14T16:46:55.471+0800: 3742.334: [GC (Allocation Failure) 2019-06-14T16:46:55.471+0800: 3742.334: [ParNew: 111908K->1571K(118016K), 0.0052707 secs] 133757K->23421K(249088K), 0.0058796 secs] [Times: user=0.02 sys=0.01, real=0.00 secs]
