+++ logs/ksql.out
@@ -1,0 +1,15 @@
+Listening for transport dt_socket at address: 5005
+SLF4J: A number (101) of logging calls during the initialization phase have been intercepted and are
+SLF4J: now being replayed. These are subject to the filtering rules of the underlying logging system.
+SLF4J: See also http://www.slf4j.org/codes.html#replay
+Jun 14, 2019 3:44:37 PM org.glassfish.jersey.internal.inject.Providers checkProviderRuntime
+WARNING: A provider io.confluent.ksql.rest.server.resources.ServerInfoResource registered in SERVER runtime does not implement any provider interfaces applicable in the SERVER runtime. Due to constraint configuration problems the provider io.confluent.ksql.rest.server.resources.ServerInfoResource will be ignored. 
+Jun 14, 2019 3:44:37 PM org.glassfish.jersey.internal.inject.Providers checkProviderRuntime
+WARNING: A provider io.confluent.ksql.rest.server.resources.StatusResource registered in SERVER runtime does not implement any provider interfaces applicable in the SERVER runtime. Due to constraint configuration problems the provider io.confluent.ksql.rest.server.resources.StatusResource will be ignored. 
+Jun 14, 2019 3:44:37 PM org.glassfish.jersey.internal.inject.Providers checkProviderRuntime
+WARNING: A provider io.confluent.ksql.rest.server.resources.KsqlResource registered in SERVER runtime does not implement any provider interfaces applicable in the SERVER runtime. Due to constraint configuration problems the provider io.confluent.ksql.rest.server.resources.KsqlResource will be ignored. 
+Jun 14, 2019 3:44:37 PM org.glassfish.jersey.internal.inject.Providers checkProviderRuntime
+WARNING: A provider io.confluent.ksql.rest.server.resources.streaming.StreamedQueryResource registered in SERVER runtime does not implement any provider interfaces applicable in the SERVER runtime. Due to constraint configuration problems the provider io.confluent.ksql.rest.server.resources.streaming.StreamedQueryResource will be ignored. 
+Jun 14, 2019 3:44:37 PM org.glassfish.jersey.internal.inject.Providers checkProviderRuntime
+WARNING: A provider io.confluent.ksql.rest.server.resources.RootDocument registered in SERVER runtime does not implement any provider interfaces applicable in the SERVER runtime. Due to constraint configuration problems the provider io.confluent.ksql.rest.server.resources.RootDocument will be ignored. 
+Listening for transport dt_socket at address: 5005
